+++ vitis/software/Baremetal/src/uz/uz_nn/uz_nn.h
@@ -40,10 +40,10 @@ void uz_nn_ff(uz_nn_t* self, uz_matrix_t const*const input);
  */
 uz_matrix_t *uz_nn_get_output_data(uz_nn_t const *const self);
 
-uz_matrix_t* uz_nn_get_bias_matrix(uz_nn_t const*const self, size_t layer);
-uz_matrix_t* uz_nn_get_weight_matrix(uz_nn_t const*const self, size_t layer);
-size_t uz_nn_get_number_of_layer(uz_nn_t const*const self);
-size_t uz_nn_get_number_of_inputs(uz_nn_t const*const self);
-size_t uz_nn_get_number_of_outputs(uz_nn_t const*const self);
+uz_matrix_t* uz_nn_get_bias_matrix(uz_nn_t const*const self, uint32_t layer);
+uz_matrix_t* uz_nn_get_weight_matrix(uz_nn_t const*const self, uint32_t layer);
+uint32_t uz_nn_get_number_of_layer(uz_nn_t const*const self);
+uint32_t uz_nn_get_number_of_inputs(uz_nn_t const*const self);
+uint32_t uz_nn_get_number_of_outputs(uz_nn_t const*const self);
 
 #endif // UZ_NN_H
